### 2024-08-21 #7126 @editorialbot
[PRE REVIEW]: GridapSolvers.jl: Scalable multiphysics finite element solvers in Julia
**Submitting author:** @JordiManyer (<a href="http://orcid.org/0000-0002-0178-3890">Jordi Manyer</a>)
**Repository:** https://github.com/gridap/GridapSolvers.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.4.0
