### 2024-08-21 #7126 @danielskatz
👋 @JordiManyer - thanks for your submission. As the track editor, I'll work on finding an editor. On your side:

* Please work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
* Also, please see the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) about how to cite multiple papers at the same time, and make changes to your .md file.
* In your references, please be sure you are not using {{}} around titles that preserve cases, as the JOSS processing determines the cases of words. If you need to preserve a the same in a word or letter, do that with just an extra internal {} where needed.  For example, `Julia` should be `{J}ulia` in one place I see.
* The paper is also a little long - it's about 1800 words, where JOSS papers are typically around 1000 max. Please see if there's anything you can shorten, perhaps by moving it to the repo/documentation and pointing to it there. You can use `@editorialbot check repository` to see how many words are in the paper.
