### 2024-08-21 #7126 @JordiManyer
Hi @danielskatz, thank you for your time. I will look into your comments and let you know if I have any doubts.
