### 2024-08-21 #7126 @danielskatz
The review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) do not require or suggest an example in the paper, so moving the example to the repo/docs would be fine.
