+++ joss_paper/paper.bib
@@ -13,11 +13,11 @@
 }
 
 @manual{mpi40,
-    author = "{Message Passing Interface Forum}",
-    title  = "{MPI}: A Message-Passing Interface Standard Version 4.0",
-    url    = "https://www.mpi-forum.org/docs/mpi-4.0/mpi40-report.pdf",
-    year   = 2021,
-    month  = jun
+  author = "{Message Passing Interface Forum}",
+  title  = "{MPI}: A Message-Passing Interface Standard Version 4.0",
+  url    = "https://www.mpi-forum.org/docs/mpi-4.0/mpi40-report.pdf",
+  year   = 2021,
+  month  = jun
 }
 
 @article{Verdugo2021,
@@ -61,7 +61,7 @@
 }
 
 @article{Bezanson2017,
-abstract = {Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia...},
+abstract = {Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. {J}ulia...},
 archivePrefix = {arXiv},
 arxivId = {1411.1607},
 author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
@@ -74,7 +74,7 @@ month = {feb},
 number = {1},
 pages = {65--98},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{Julia: a fresh approach to numerical computing}},
+title = {{J}ulia: a fresh approach to numerical computing},
 volume = {59},
 year = {2017}
 }
@@ -89,7 +89,7 @@ month = {aug},
 number = {52},
 pages = {2520},
 publisher = {The Open Journal},
-title = {{Gridap: an extensible finite element toolbox in Julia}},
+title = {{G}ridap: an extensible finite element toolbox in {J}ulia},
 url = {https://joss.theoj.org/papers/10.21105/joss.02520},
 volume = {5},
 year = {2020}
@@ -108,7 +108,7 @@ month = {dec},
 number = {6},
 pages = {C436--C468},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{A generic finite element framework on parallel tree-based adaptive meshes}},
+title = {A generic finite element framework on parallel tree-based adaptive meshes},
 volume = {42},
 year = {2020}
 }
@@ -124,7 +124,7 @@ month = {may},
 number = {3},
 pages = {1103--1133},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{p4est: scalable algorithms for parallel adaptive mesh refinement on forests of octrees}},
+title = {{p4est}: scalable algorithms for parallel adaptive mesh refinement on forests of octrees},
 volume = {33},
 year = {2011}
 }
@@ -153,7 +153,7 @@ year = {2011}
 @Article{libMeshPaper,
   doi     = {10.1007/s00366-006-0049-3},
   author  = {B.~S.~Kirk and J.~W.~Peterson and R.~H.~Stogner and G.~F.~Carey},
-  title   = {{\texttt{libMesh}: A C++ library for parallel adaptive mesh refinement/coarsening simulations}},
+  title   = {{\texttt{libMesh}}: A {C}++ library for parallel adaptive mesh refinement/coarsening simulations},
   journal = {Engineering with Computers},
   volume  = 22,
   number  = {3--4},
@@ -192,7 +192,7 @@ journal = {ACM Transactions on Mathematical Software},
[…]
