### 2024-08-21 #7126 @JordiManyer
I think the DOIs missing at this point are just websites and misc references.
