+++ test/_dev/Vanka/stokesdarcy_vanka_GMG.jl
@@ -1,3 +1,4 @@
+
 using Gridap
 using GridapSolvers
 using LinearAlgebra
@@ -15,31 +16,59 @@ function l2_error(uh,u_exact,dΩ)
   return sqrt(sum(∫(eh⋅eh)dΩ))
 end
 
-u_exact(x) = VectorValue(-x[1],x[2])
-p_exact(x) = x[1] - 0.5
+function add_labels_2d(model)
+  labels = get_face_labeling(model)
+  add_tag_from_tags!(labels,"newman",[5])
+  add_tag_from_tags!(labels,"dirichlet",[collect(1:4)...,6,7,8])
+end
 
-Dc = 2
-cmodel = CartesianDiscreteModel((0,1,0,1),(4,4))
-labels = get_face_labeling(cmodel)
-add_tag_from_tags!(labels,"newman",[5])
-add_tag_from_tags!(labels,"dirichlet",[collect(1:4)...,6,7,8])
+function add_labels_3d(model)
+  labels = get_face_labeling(model)
+  add_tag_from_tags!(labels,"newman",[21])
+  add_tag_from_tags!(labels,"dirichlet",[collect(1:20)...,22,23,24,25,26])
+end
+
+B = VectorValue(0.0,0.0,1.0)
+u_exact(x) = VectorValue(x[1],-x[2],0.0)
+p_exact(x) = sum(x)
+j_exact(x) = VectorValue(-x[2],-x[1],0.0) + VectorValue(1.0,1.0,0.0)
+φ_exact(x) = x[1] + x[2]
+
+Dc = 3
+domain = (Dc == 2) ? (0,1,0,1) : (0,1,0,1,0,1)
+ncells = (Dc == 2) ? (4,4) : (2,2,2)
+cmodel = CartesianDiscreteModel(domain,ncells)
+if Dc == 2
+  add_labels_2d(cmodel)
+else
+  add_labels_3d(cmodel)
+end
 model = refine(cmodel,2)
 
 order = 2
-reffe_u = ReferenceFE(lagrangian,VectorValue{2,Float64},order)
+reffe_u = ReferenceFE(lagrangian,VectorValue{Dc,Float64},order)
 reffe_p = ReferenceFE(lagrangian,Float64,order-1;space=:P)
+reffe_j = ReferenceFE(raviart_thomas,Float64,order-1)
+reffe_φ = ReferenceFE(lagrangian,Float64,order-1)
 
 VH = TestFESpace(cmodel,reffe_u;dirichlet_tags="dirichlet")
 UH = TrialFESpace(VH,u_exact)
 QH = TestFESpace(cmodel,reffe_p,conformity=:L2)
-XH = MultiFieldFESpace([UH,QH])
-YH = MultiFieldFESpace([VH,QH])
+DH = TestFESpace(cmodel,reffe_j;dirichlet_tags="dirichlet")
+JH = TrialFESpace(DH,j_exact)
+ΦH = TestFESpace(cmodel,reffe_φ,conformity=:L2)
+
+XH = MultiFieldFESpace([UH,QH,JH,ΦH])
+YH = MultiFieldFESpace([VH,QH,DH,ΦH])
 
 Vh = TestFESpace(model,reffe_u;dirichlet_tags="dirichlet")
 Uh = TrialFESpace(Vh,u_exact)
[…]
