+++ test/_dev/Vanka/darcy_vanka_GMG.jl
@@ -57,8 +57,8 @@ f(x) = u_exact(x) - ∇(p_exact)(x)
 σ(x) = p_exact(x)
 graddiv(u,v,dΩ) = ∫((∇⋅u)⋅(∇⋅v))*dΩ
 function a((u,p),(v,q),dΩ)
-  c = ∫(u⋅v + (∇⋅v)*p - (∇⋅u)*q)dΩ
-  if !iszero(α)
+  c = ∫(u⋅v + (∇⋅v)*p + (∇⋅u)*q)dΩ
+  if α > 0.0
     c += graddiv(u,v,dΩ)
   end
   return c
@@ -82,8 +82,9 @@ l2_error(uh_exact,u_exact,dΩh)
 l2_error(ph_exact,p_exact,dΩh)
 l2_norm(∇⋅uh_exact,dΩh)
 
+# NOTE: Convergence seem quite dependent on the damping weight. For ω=0.2, we diverge...
 PD = PatchDecomposition(model)
-smoother = RichardsonSmoother(VankaSolver(Xh,PD),5,0.2)
+smoother = RichardsonSmoother(VankaSolver(Xh,PD),10,0.1)
 smoother_ns = numerical_setup(symbolic_setup(smoother,Ah),Ah)
 
 function project_f2c(rh)
