### 2024-08-29 56e9821
GMG working in serial
