+++ test/_dev/Vanka/mhd_uj_vanka_GMG.jl
@@ -51,24 +51,24 @@ reffe_p = ReferenceFE(lagrangian,Float64,order-1;space=:P)
 reffe_j = ReferenceFE(raviart_thomas,Float64,order-1)
 reffe_φ = ReferenceFE(lagrangian,Float64,order-1)
 
+using Badia2024
+using SpecialPolynomials
+using Polynomials
+reffe_j = Badia2024.SpecialRTReffe(ChebyshevT,order,Dc)
+
 VH = TestFESpace(cmodel,reffe_u;dirichlet_tags="dirichlet")
 UH = TrialFESpace(VH,u_exact)
-QH = TestFESpace(cmodel,reffe_p,conformity=:L2)
 DH = TestFESpace(cmodel,reffe_j;dirichlet_tags="dirichlet")
 JH = TrialFESpace(DH,j_exact)
-ΦH = TestFESpace(cmodel,reffe_φ,conformity=:L2)
-
-XH = MultiFieldFESpace([UH,QH,JH,ΦH])
-YH = MultiFieldFESpace([VH,QH,DH,ΦH])
+XH = MultiFieldFESpace([UH,JH])
+YH = MultiFieldFESpace([VH,DH])
 
 Vh = TestFESpace(model,reffe_u;dirichlet_tags="dirichlet")
 Uh = TrialFESpace(Vh,u_exact)
-Qh = TestFESpace(model,reffe_p,conformity=:L2)
 Dh = TestFESpace(model,reffe_j;dirichlet_tags="dirichlet")
 Jh = TrialFESpace(Dh,j_exact)
-Φh = TestFESpace(model,reffe_φ,conformity=:L2)
-Xh = MultiFieldFESpace([Uh,Qh,Jh,Φh])
-Yh = MultiFieldFESpace([Vh,Qh,Dh,Φh])
+Xh = MultiFieldFESpace([Uh,Jh])
+Yh = MultiFieldFESpace([Vh,Dh])
 
 qdegree = 2*order
 Ωh = Triangulation(model)
@@ -81,27 +81,32 @@ dΩHh = Measure(ΩH,Ωh,qdegree)
 dΓh = Measure(Γh,qdegree)
 n = get_normal_vector(Γh)
 
-ν = 1.e-2
-α = -1.0
-I_tensor = one(TensorValue{Dc,Dc,Float64})
-f(x) = -Δ(u_exact)(x) - ∇(p_exact)(x) - cross(j_exact(x),B)
-g(x) = j_exact(x) - ∇(φ_exact)(x) - cross(u_exact(x),B)
-σ(x) = ∇(u_exact)(x) + p_exact(x)*I_tensor
-γ(x) = φ_exact(x)*I_tensor
+ν = 1.e-8
+α = 1.0
+f(x) = -(-Δ(u_exact)(x) - cross(j_exact(x),B))
+g(x) = j_exact(x) - cross(u_exact(x),B)
+σ(x) = ∇(u_exact)(x)
+
 crossB(u,v,dΩ) = ∫(cross(u,B)⋅v)dΩ
 Π = GridapSolvers.MultilevelTools.LocalProjectionMap(divergence,reffe_p,qdegree)
 Πgraddiv(u,v,dΩ) = ∫(α*Π(u)⋅(∇⋅v))*dΩ
 graddiv(u,v,dΩ) = ∫(α*(∇⋅u)⋅(∇⋅v))*dΩ
-function a((u,p,j,φ),(v,q,d,ψ),dΩ)
-  c = ∫(ν*∇(u)⊙∇(v) + (∇⋅v)*p - (∇⋅u)*q)dΩ
-  c = c + ∫(j⋅d + (∇⋅d)*φ - (∇⋅j)*ψ)dΩ
[…]
