+++ src/PatchBasedSmoothers/seq/PatchBasedLinearSolvers.jl
@@ -73,112 +73,92 @@ struct PatchBasedSmootherNumericalSetup{A,B,C,D} <: Gridap.Algebra.NumericalSetu
   caches   :: D
 end
 
+function assemble_patch_matrices(Ph::PatchFESpace,ap;local_solver=LUSolver())
+  assem = SparseMatrixAssembler(Ph,Ph)
+  Ap    = assemble_matrix(ap,assem,Ph,Ph)
+  Ap_ns = numerical_setup(symbolic_setup(local_solver,Ap),Ap)
+  return Ap, Ap_ns
+end
+
+function assemble_patch_matrices(Ph::DistributedPatchFESpace,ap;local_solver=LUSolver())
+  u, v  = get_trial_fe_basis(Vh), get_fe_basis(Vh)
+  matdata = collect_cell_matrix(Ph,Ph,ap(u,v))
+  Ap, Ap_ns = map(local_views(Ph),matdata) do Ph, matdata
+    assem = SparseMatrixAssembler(Ph,Ph)
+    Ap    = assemble_matrix(assem,matdata)
+    Ap_ns = numerical_setup(symbolic_setup(local_solver,Ap),Ap)
+    return Ap, Ap_ns
+  end |> PartitionedArrays.tuple_of_arrays
+  return Ap, Ap_ns
+end
+
+function update_patch_matrices!(Ap,Ap_ns,Ph::PatchFESpace,ap)
+  assem = SparseMatrixAssembler(Ph,Ph)
+  assemble_matrix!(Ap,assem,Ph,Ph,ap)
+  numerical_setup!(Ap_ns,Ap)
+end
+
+function update_patch_matrices!(Ap,Ap_ns,Ph::DistributedPatchFESpace,ap)
+  u, v  = get_trial_fe_basis(Vh), get_fe_basis(Vh)
+  matdata = collect_cell_matrix(Ph,Ph,ap(u,v))
+  map(Ap, Ap_ns, local_views(Ph), matdata) do Ap, Ap_ns, Ph, matdata
+    assem = SparseMatrixAssembler(Ph,Ph)
+    assemble_matrix!(Ap,assem,matdata)
+    numerical_setup!(Ap_ns,Ap)
+  end
+end
+
+function allocate_patch_workvectors(Ph::PatchFESpace,Vh::FESpace)
+  rp     = zero_free_values(Ph)
+  dxp    = zero_free_values(Ph)
+  return rp,dxp
+end
+
+function allocate_patch_workvectors(Ph::DistributedPatchFESpace,Vh::GridapDistributed.DistributedFESpace)
+  rp     = zero_free_values(Ph)
+  dxp    = zero_free_values(Ph)
+  r      = zero_free_values(Vh)
+  x      = zero_free_values(Vh)
+  return rp,dxp,r,x
+end
+
 function Gridap.Algebra.numerical_setup(ss::PatchBasedSymbolicSetup,A::AbstractMatrix)
   @check !ss.solver.is_nonlinear
   solver = ss.solver
+  local_solver = solver.local_solver
   Ph, Vh = solver.patch_space, solver.space
-  weights = solver.weighted ? compute_weight_operators(Ph,Vh) : nothing
   
   ap(u,v) = solver.biform(u,v)
-  assem = SparseMatrixAssembler(Ph,Ph)
-  Ap    = assemble_matrix(ap,assem,Ph,Ph)
-  Ap_ns = numerical_setup(symbolic_setup(solver.local_solver,Ap),Ap)
-
-  # Caches
-  rp     = allocate_in_range(Ap); fill!(rp,0.0)
-  dxp    = allocate_in_domain(Ap); fill!(dxp,0.0)
-  caches = (rp,dxp)
-
[…]
