+++ test/LinearSolvers/SmoothersTests.jl
@@ -5,7 +5,6 @@ using MPI
 using Gridap, Gridap.Algebra
 using GridapDistributed
 using PartitionedArrays
-using IterativeSolvers
 
 using GridapSolvers
 using GridapSolvers.LinearSolvers
@@ -29,15 +28,10 @@ function smoothers_driver(parts,model,P)
   op = AffineFEOperator(a,l,Uh,Vh)
   A, b = get_matrix(op), get_vector(op)
 
-  ss = symbolic_setup(P,A)
-  ns = numerical_setup(ss,A)
-
-  x = allocate_in_domain(A); ; fill!(x,zero(eltype(x)))
-  x, history = IterativeSolvers.cg!(x,A,b;
-                                    verbose=i_am_main(parts),
-                                    reltol=1.0e-8,
-                                    Pl=ns,
-                                    log=true)
+  solver = CGSolver(P;rtol=1.0e-8,verbose=i_am_main(parts))
+  ns = numerical_setup(sumbolic_setup(solver,A),A)
+  x = allocate_in_domain(A); fill!(x,zero(eltype(x)))
+  solve!(x,ns,b)
 
   u  = interpolate(sol,Uh)
   uh = FEFunction(Uh,x)

[…]
