### 2024-09-04 fa5f2d7
Added patch closures
