+++ src/PatchBasedSmoothers/seq/PatchDecompositions.jl
@@ -23,21 +23,17 @@ of `Ω` such that `Ω = Σ_i Ω_i`.
 - `Dr::Integer` : Dimension of the patch root
 - `model::DiscreteModel{Dc,Dp}` : Underlying discrete model
 - `patch_cells::Table` : [patch][local cell] -> cell
-- `patch_cells_overlapped::Table`        : [patch][local cell] -> overlapped cell
 - `patch_cells_faces_on_boundary::Table` : [d][overlapped cell][local face] -> face is on patch boundary
 
 """
 struct PatchDecomposition{Dr,Dc,Dp} <: GridapType
   model                         :: DiscreteModel{Dc,Dp}
-  patch_cells                   :: Gridap.Arrays.Table         # [patch][local cell] -> cell
-  patch_cells_overlapped        :: Gridap.Arrays.Table         # [patch][local cell] -> overlapped cell
-  patch_cells_faces_on_boundary :: Vector{Gridap.Arrays.Table} # [d][overlapped cell][local face] -> face is on patch boundary
+  patch_cells                   :: Arrays.Table         # [patch][local cell] -> cell
+  patch_cells_faces_on_boundary :: Vector{Arrays.Table} # [d][overlapped cell][local face] -> face is on patch boundary
+  patch_boundary_style          :: PatchBoundaryStyle
 end
 
-num_patches(a::PatchDecomposition) = length(a.patch_cells)
-Gridap.Geometry.num_cells(a::PatchDecomposition) = length(a.patch_cells.data)
-
-@doc """
+"""
     function PatchDecomposition(
       model::DiscreteModel{Dc,Dp};
       Dr=0,
@@ -49,31 +45,112 @@ Returns an instance of [`PatchDecomposition`](@ref) from a given discrete model.
 """
 function PatchDecomposition(
   model::DiscreteModel{Dc,Dp};
-  Dr=0,
+  Dr::Integer=0,
   patch_boundary_style::PatchBoundaryStyle=PatchBoundaryExclude(),
   boundary_tag_names::AbstractArray{String}=["boundary"]
 ) where {Dc,Dp}
-  Gridap.Helpers.@check 0 <= Dr <= Dc-1
+  @assert 0 <= Dr <= Dc-1
 
   topology     = get_grid_topology(model)
-  patch_cells  = Gridap.Geometry.get_faces(topology,Dr,Dc)
-  patch_facets = Gridap.Geometry.get_faces(topology,Dr,Dc-1)
-  patch_cells_overlapped = compute_patch_overlapped_cells(patch_cells)
+  patch_cells  = Geometry.get_faces(topology,Dr,Dc)
+  patch_facets = Geometry.get_faces(topology,Dr,Dc-1)
 
   patch_cells_faces_on_boundary = compute_patch_cells_faces_on_boundary(
-    model, patch_cells, patch_cells_overlapped,
-    patch_facets, patch_boundary_style, boundary_tag_names
+    model, patch_cells, patch_facets, patch_boundary_style, boundary_tag_names
   )
 
   return PatchDecomposition{Dr,Dc,Dp}(
-    model, patch_cells, patch_cells_overlapped, patch_cells_faces_on_boundary
+    model, patch_cells, patch_cells_faces_on_boundary, patch_boundary_style
   )
 end
 
-function compute_patch_overlapped_cells(patch_cells)
-  num_overlapped_cells = length(patch_cells.data)
-  data = Gridap.Arrays.IdentityVector(num_overlapped_cells)
-  return Gridap.Arrays.Table(data,patch_cells.ptrs)
+"""
+    num_patches(a::PatchDecomposition)
+"""
[…]

[…]
