### 2024-09-04 651f330
Added set_depth!
