### 2024-09-06 d137360
Resolved minor issues when changing triangulations
