+++ src/PatchBasedSmoothers/CoarsePatchDecompositions.jl
@@ -11,8 +11,7 @@ function CoarsePatchDecomposition(
   patch_cells_overlapped = compute_patch_overlapped_cells(patch_cells)
   patch_facets = get_coarse_patch_facets(model, patch_cells)
   patch_cells_faces_on_boundary = compute_patch_cells_faces_on_boundary(
-    model, patch_cells, patch_cells_overlapped,
-    patch_facets, patch_boundary_style, boundary_tag_names
+    model, patch_cells, patch_facets, patch_boundary_style, boundary_tag_names
   )
 
   return PatchDecomposition{Dc,Dc,Dp}(
