+++ src/PatchBasedSmoothers/DistributedPatchFESpaces.jl
@@ -1,4 +1,5 @@
 
+
 const DistributedPatchFESpace = GridapDistributed.DistributedSingleFieldFESpace{<:AbstractVector{<:PatchFESpace}}
 
 function PatchFESpace(
