### 2024-09-18 fac43c2
patch-wise patch solvers
