+++ src/PatchBasedSmoothers/PatchFESpaces.jl
@@ -106,11 +106,10 @@ function PatchFESpace(
   cell_dofs_ids = get_cell_dof_ids(space)
   patch_cells_overlapped = get_patch_cells_overlapped(patch_decomposition)
   patch_cell_dofs_ids, num_dofs = generate_patch_cell_dofs_ids(
-    get_grid_topology(patch_decomposition.model),
     patch_decomposition.patch_cells,
     patch_cells_overlapped,
     patch_decomposition.patch_cells_faces_on_boundary,
-    cell_dofs_ids,cell_conformity,patches_mask
+    cell_dofs_ids,cell_conformity;patches_mask
   )
   dof_to_pdof = generate_dof_to_pdof(space,patch_decomposition,patch_cell_dofs_ids)
   return PatchFESpace(space,patch_decomposition,num_dofs,patch_cell_dofs_ids,dof_to_pdof)
@@ -182,72 +181,71 @@ end
 # Construction of the patch cell dofs ids
 
 function generate_patch_cell_dofs_ids(
-  topology,
   patch_cells,
   patch_cells_overlapped,
   patch_cells_faces_on_boundary,
   cell_dofs_ids,
-  cell_conformity,
-  patches_mask
+  cell_conformity::CellConformity;
+  patches_mask = Fill(false,length(patch_cells)),
+  numbering = :global
 )
+  @assert numbering in [:global,:local]
   patch_cell_dofs_ids = allocate_patch_cell_array(patch_cells,cell_dofs_ids;init=Int32(-1))
   num_dofs = generate_patch_cell_dofs_ids!(
-    patch_cell_dofs_ids,topology,
+    patch_cell_dofs_ids,
     patch_cells,patch_cells_overlapped,
     patch_cells_faces_on_boundary,
-    cell_dofs_ids,cell_conformity,patches_mask
+    cell_dofs_ids,cell_conformity;
+    patches_mask,numbering
   )
   return patch_cell_dofs_ids, num_dofs
 end
 
 function generate_patch_cell_dofs_ids!(
   patch_cell_dofs_ids,
-  topology,
   patch_cells,
   patch_cells_overlapped,
   patch_cells_faces_on_boundary,
   cell_dofs_ids,
-  cell_conformity,
-  patches_mask
+  cell_conformity::CellConformity;
+  patches_mask = Fill(false,length(patch_cells)),
+  numbering = :global
 )
-  cache = array_cache(patch_cells)
-  num_patches = length(patch_cells)
-  current_dof = 1
-  for patch = 1:num_patches
-    current_patch_cells = getindex!(cache,patch_cells,patch)
-    current_dof = generate_patch_cell_dofs_ids!(
+  @assert numbering in [:global,:local]
+  dof_offset = 1
+  for patch = 1:length(patch_cells)
+    current_patch_cells = view(patch_cells,patch)
+    dof_offset = generate_patch_cell_dofs_ids!(
       patch_cell_dofs_ids,
-      topology,
-      patch,
-      current_patch_cells,
-      patch_cells_overlapped,
+      patch,current_patch_cells,patch_cells_overlapped,
       patch_cells_faces_on_boundary,
-      cell_dofs_ids,
-      cell_conformity;
-      free_dofs_offset=current_dof,
+      cell_dofs_ids,cell_conformity;
+      dof_offset=dof_offset,
       mask=patches_mask[patch]
     )
+    if numbering == :local
+      dof_offset = 1
+    end
   end
-  return current_dof-1
+  return dof_offset-1
 end
 
-# Note: we do not actually need to generate a global numbering for Dirichlet DoFs. We can
[…]
