+++ joss_paper/paper.bib
@@ -294,16 +294,16 @@ eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9780898719505}
 } 
 
 @inproceedings{Arnold1,
-  title={Preconditing in H(div) and applications},
-  author={D. N. Arnold and Richard S. Falk and Ragnar Winther},
+  title={Preconditing in {H}(div) and applications},
+  author={D. N. Arnold and R. S. Falk and R. Winther},
   year={1997},
   url={https://api.semanticscholar.org/CorpusID:12559456},
   doi={10.1090/S0025-5718-97-00826-0}
 }
 
 @article{Arnold2,
-  title={Multigrid in H(div) and H(curl)},
-  author={D. N. Arnold and Richard S. Falk and Ragnar Winther},
+  title={Multigrid in {H}(div) and {H}(curl)},
+  author={D. N. Arnold and R. S. Falk and R. Winther},
   journal={Numerische Mathematik},
   year={2000},
   volume={85},
