### 2024-09-25 #7162 @Leticia-maria
@diehlpk I finished my checklist, can I also leave some comments here?
- Review reference syntax
