### 2024-09-25 4813c19
LocalProjectionMap bugfix in parallel
