### 2024-09-25 #7162 @diehlpk
> @diehlpk I finished my checklist, can I also leave some comments here?
> - Review reference syntax
> - More details on how the benchmark was performed: details about which benchmark method was used. It was not clear if the algorithm is compute-bound or memory-bound, maybe expand a little bit more on that.
> - Overall, good propositions, I think it fits into the journal, with some corrections.

@JordiManyer please work on these comments. After that I will look at the paper.
