### 2024-09-26 #7162 @JordiManyer
> Review reference syntax

The skipped DOIs are for references that link to repositories or code papers. I tried my best to find suitable DOIs for them, but I couldn't. It's usually documents that are published by developers but not in journals (like manuals, etc..)

> More details on how the benchmark was performed

I can expand on this, but just FYI the paper was already flagged for being too long.

> It was not clear if the algorithm is compute-bound or memory-bound

It's quite hard to tell. It's not an easy algorithm, and I don't think I've ever seen anyone touch on this in a FEM paper.
Krylov algorithms are memory-bound due to most of their cost being associated to sparse matrix-vector products. However, the picture changes when adding a complex preconditioner (which usually takes most of the cost). In our case, the block-preconditioner is definitely taking most of the cost.

The pressure solver is quite cheap (basically a smaller Krylov-solved solver), but the GMG used for the velocity block is not.

Within the GMG, there are multiple ingredients that can affect performance:
- The smoother, which is this case is a Richardson iterator with Jacobi smoother. Again, most of the cost is associated to a sparse matrix-vector product and therefore memory bound.
- The restriction and prolongation operators, which solve L2 projections involving integrals over the global domain and another Krylov solver.
- The coarse solver, which is a LU backwards-forwards substitution (which should also be memory bound).

This is not even taking into account that we are in an MPI setting with hundreds of processors, where overhead due to communications cannot be neglected.

I could spend another 500 words arguing this in the paper, but I don't think it's within the scope. Specially in a FEM paper, where the cost per iteration is not important compared to the gains you obtain by using a good preconditioner (in terms of number of iterations).

Maybe @amartinhuertas can give his opinion as well.
