### 2024-09-26 #7162 @amartinhuertas
I agree with @JordiManyer insights.

Long history short: preconditioned iterative solvers for large-scale PDEs are mostly memory-bound (when one uses sparse matrix storage as in the paper/current codes).
