### 2024-09-26 #7162 @Leticia-maria
I think that clarifies all the questions on my end! The paper should be good to go.
