### 2024-09-30 d4be460
Fixed PatchSkeletons for adapted models
