+++ test/_dev/PatchBased/BoundaryTrians.jl
@@ -14,6 +14,12 @@ using GridapSolvers.LinearSolvers
 using GridapSolvers.MultilevelTools
 using GridapSolvers.PatchBasedSmoothers
 
+function get_cell_size(Ω::Triangulation{Dc}) where Dc
+  dΩ = Measure(Ω,1)
+  h_e = Gridap.CellData.get_array(∫(1)dΩ) .^ (1/Dc)
+  return CellField(h_e,Ω)
+end
+
 function weakforms(model)
   Ω = Triangulation(model)
   Γ = BoundaryTriangulation(model)
@@ -26,10 +32,13 @@ function weakforms(model)
   n_Γ = get_normal_vector(Γ)
   n_Λ = get_normal_vector(Λ)
 
+  h_e = get_cell_size(Λ)
+
   a1(u,v) = ∫(∇(u)⊙∇(v))dΩ
   a2(u,v) = ∫((∇(v)⋅n_Γ)⋅u)dΓ
   a3(u,v) = ∫(jump(u⋅n_Λ)⋅jump(v⋅n_Λ))dΛ
-  return a1, a2, a3
+  a4(u,v) = ∫(h_e*jump(u⋅n_Λ)⋅jump(v⋅n_Λ))dΛ
+  return a1, a2, a3, a4
 end
 
 model = CartesianDiscreteModel((0,1,0,1),(2,2))
@@ -52,8 +61,8 @@ Ph = PatchFESpace(Vh,PD,reffe)
 
 Γp_virtual = BoundaryTriangulation(PD,tags=["boundary","interior"],reverse=true)
 
-a1, a2, a3 = weakforms(model)
-ap1, ap2, ap3 = weakforms(PD)
+a1, a2, a3, a4 = weakforms(model)
+ap1, ap2, ap3, ap4 = weakforms(PD)
 
 A1 = assemble_matrix(a1,Vh,Vh)
 Ap1 = assemble_matrix(ap1,Ph,Ph)
@@ -63,3 +72,10 @@ Ap2 = assemble_matrix(ap2,Ph,Ph)
 
 A3 = assemble_matrix(a3,Vh,Vh)
 Ap3 = assemble_matrix(ap3,Ph,Ph)
+
+A4 = assemble_matrix(a4,Vh,Vh)
+Ap4 = assemble_matrix(ap4,Ph,Ph)
+
+
+
+
