### 2024-10-02 #7162 @JordiManyer
My checks:
