### 2024-10-02 #7162 @diehlpk
@JordiManyer the title of the Zenodo archive is

gridap/GridapSolvers.jl: v0.4.1

but it should be the same as of the paper. There is no need to obtain a new DOI and you can just edit the title.
