### 2024-10-04 #7162 @danielskatz
@JordiManyer - as AEiC, I've proofread your paper and suggest the changes in https://github.com/gridap/GridapSolvers.jl/pull/75 - please merge this, or let know what you disagree with, then we can proceed to acceptance and publication.
