### 2024-10-05 #7162 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
