### 2024-10-05 #7162 @danielskatz
Congratulations to @JordiManyer (Jordi Manyer) and co-authors on your publication!!

JOSS depends on volunteers and couldn't be successful without you
